+++ lib/segmented_progress_bar.dart
@@ -1,4 +1,2 @@
-library segmented_progress_bar;
-
 export 'src/models/types.dart';
 export 'src/segmented_progress_bar.dart';